requirements.txt
matplotlib
pillow
pandas
numpy
tensorflow[and-cuda]
jupyterlab
opencv-python
seaborn